Move layout debug flag to Radiant rebuild
